requirements.txt
streamlit==1.13.0
whoosh==2.7.4
PyPDF2==1.26.0
python-docx==0.8.11
requests==2.28.1
beautifulsoup4==4.11.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
